repo improved,scripts added

+++ contracts/BurnAuction.sol
@@ -1,212 +1,281 @@
 pragma solidity ^0.5.15;
 
 contract BurnAuction {
-    
-	// 1 next 1.3 bonus 10 1.3 10 0.13  1.3-1-0.13=burned
-	// discuss if you want to incentivize early bidders for slots using bonus skip for now
-	// todo make uint types explicit
-	// discuss need for default Coordinator-keep it
-	// discuss need to prevent submit batches multiple times after knowing hubble architecture
-		
-	// number of blocks available in slot
-	uint32 public blocksPerSlot;
-	// number of blocks after contract deployment for genesisBlock number
-	uint public delayGenesis;
-	// First block where the first slot begins
-    uint public genesisBlock;
-	// Maximum rollup transactions: either off-chain or on-chain transactions
-    uint public maxTx;
-	// Min differance between currentslot and auction slots
-	uint public minNextSlots;
-	// Burn Address
+    // 1 next 1.3 bonus 10 1.3 10 0.13  1.3-1-0.13=burned
+    // discuss if you want to incentivize early bidders for slots using bonus-skip for now
+    // discuss need for default Coordinator-keep it-done
+    // discuss need to prevent submit batches multiple times after knowing hubble architecture--handled in rollup.sol if required
+    // todo use safemath
+    // todo make uint types explicit
+
+    // number of blocks available in slot
+    uint32 public blocksPerSlot;
+    // number of blocks after contract deployment for genesisBlock number
+    uint256 public delayGenesis;
+    // First block where the first slot begins
+    uint256 public genesisBlock;
+    // Maximum rollup transactions: either off-chain or on-chain transactions
+    uint256 public maxTx;
+    // Min differance between currentslot and auction slots
+    uint256 public minNextSlots;
+    // Burn Address
     address payable burnAddress;
-	// Minimum bid to enter the auction
-	uint public minBid;
+    // Minimum bid to enter the auction
+    uint256 public minBid;
     // Default Coordinator
-	// will have to decide if we want to keep default Coordinator or not
+    // will have to decide if we want to keep default Coordinator or not
     Coordinator public coDefault;
-	
-	// Coordinator structure
+
+    // Coordinator structure
     struct Coordinator {
-	    // address to return unsuccessful bid funds back
+        // address to return unsuccessful bid funds back
         address payable returnAddress;
-		// Coordinator address having right to submit batch
+        // Coordinator address having right to submit batch
         address submitBatchAddress;
-		// Coordinator url 
+        // Coordinator url
         string url;
     }
 
     // bid structure
     struct Bid {
-	    // bid amount = sumtotalFees + targetProfit
-        uint amount;
-		// used to indicate active auction for slot
+        // bid amount = sumtotalFees + targetProfit
+        uint256 amount;
+        // used to indicate active auction for slot
         bool initialized;
-		// profit per auction slot by Coordinator
-        uint targetProfit;
-	    // sum of fees of all transactions included in slot
-	    uint sumtotalFees;
+        // profit per auction slot by Coordinator
+        uint256 targetProfit;
+        // sum of fees of all transactions included in slot
+        uint256 sumtotalFees;
     }
-	
+
     //might be removed
-	// information of slot structure
+    // information of slot structure
     struct InfoSlot {
         // current price of slot
-        uint slotPrice;
+        uint256 slotPrice;
         // current max target profit on slot
-		uint maxtargetProfit;
+        uint256 maxtargetProfit;
     }
-	
-	// Mappings
-	// mapping to control winner of slot
-    mapping(uint => Coordinator) public slotWinner;
+
+    // Mappings
+    // mapping to control winner of slot
+    mapping(uint256 => Coordinator) public slotWinner;
     // mapping to control bid by slot
-    mapping(uint => Bid) public slotBid;
+    mapping(uint256 => Bid) public slotBid;
     // mapping to control information of slot
-    mapping(uint => InfoSlot) public infoSlot;
-	
-	// events
-	/**
+    mapping(uint256 => InfoSlot) public infoSlot;
+
+    // events
+    /**
      * @dev Event called when an Coordinator beat the current best bid of an ongoing auction
      */
-    event currentBestBid(uint32 slot, uint amount, uint sumtotalFees, uint targetProfit, address Coordinator, string url);
-    
-	// todo move all governance parameters in Constructor-done
-	/**
+    event currentBestBid(
+        uint32 slot,
+        uint256 amount,
+        uint256 sumtotalFees,
+        uint256 targetProfit,
+        address Coordinator,
+        string url
+    );
+
+    // todo move all governance parameters in Constructor-done
+    /**
      * @dev BurnAuction Constructor
      * Set first block where the first slot begin
      * @param _maxTx maximum transactions
      * @param _burnAddress burner address
-	 * @param _blocksPerSlot number of blocks in slot
-	 * @param _delayGenesis delay genesisBlock by this
-	 * @param _minBid minimum bid for a auction slot
-	 * @param _minNextSlots differance between currentslot and auction slot
+     * @param _blocksPerSlot number of blocks in slot
+     * @param _delayGenesis delay genesisBlock by this
+     * @param _minBid minimum bid for a auction slot
+     * @param _minNextSlots differance between currentslot and auction slot
      */
     constructor(
-	    uint256 _maxTx,
-	    address payable _burnAddress,
-	    uint32 _blocksPerSlot,
-		uint _delayGenesis,
-		uint _minBid,
-		uint _minNextSlots,
-		address payable coReturnAddress,
-		address coSubmitBatchAddress,
-		string memory coUrl
-	) public {
+        uint256 _maxTx,
+        address payable _burnAddress,
+        uint32 _blocksPerSlot,
+        uint256 _delayGenesis,
+        uint256 _minBid,
+        uint256 _minNextSlots,
+        address payable coReturnAddress,
+        address coSubmitBatchAddress,
+        string memory coUrl
+    ) public {
         genesisBlock = getBlockNumber() + _delayGenesis;
         maxTx = _maxTx;
         burnAddress = _burnAddress;
-		blocksPerSlot = _blocksPerSlot;
-		delayGenesis = _delayGenesis;
-		minBid = _minBid;
-		minNextSlots = _minNextSlots;		
-		coDefault = Coordinator(coReturnAddress,coSubmitBatchAddress,coUrl);
+        blocksPerSlot = _blocksPerSlot;
+        delayGenesis = _delayGenesis;
+        minBid = _minBid;
+        minNextSlots = _minNextSlots;
+        coDefault = Coordinator(coReturnAddress, coSubmitBatchAddress, coUrl);
     }
-	
-	// functions
-	//update with natspec comments
-	
-	function bid(
-	    uint32 slot,
-		Coordinator memory co,
-		uint _sumtotalFees,
-		uint _targetProfit,
-		uint value
-	) internal returns (uint) {
-	    uint burnBid = 0;
-		uint amount = _sumtotalFees + _targetProfit;
-		//require checks
-		require(value >= amount,"Ether sent not enough");
-		if(slotBid[slot].initialized) {
-		require(_sumtotalFees >= slotBid[slot].sumtotalFees,"include more txns");
-		require(_targetProfit < slotBid[slot].targetProfit,"take less profits");
-        // should we refund previous bidder?
-		} else {
-		//amount greater than minBid
-		require(amount >= minBid,"bid not enough than minimum bid");		
-		slotBid[slot].initialized = true;
-		}
-		slotWinner[slot] = co;
-		// update bid info for slot
+
+    // functions
+    // todo update with natspec comments
+
+    // todo improve this logic,fix bugs
+    //pricing logic needs improvement,will come back here after writing few tests
+    function bid(
+        uint32 slot,
+        Coordinator memory co,
+        uint256 _sumtotalFees,
+        uint256 _targetProfit,
+        uint256 value
+    ) internal returns (uint256) {
+        uint256 burnBid = 0;
+        uint256 amount = _sumtotalFees + _targetProfit;
+        //require checks
+        require(value >= amount, "Ether sent not enough");
+        if (slotBid[slot].initialized) {
+            require(
+                _sumtotalFees >= slotBid[slot].sumtotalFees,
+                "include more txns"
+            );
+            require(
+                _targetProfit < slotBid[slot].targetProfit,
+                "take less profits"
+            );
+            // should we refund previous bidder?
+            //problem as burnbid = sumofallfees-targetprofit as in spec
+        } else {
+            //amount greater than minBid
+            require(amount >= minBid, "bid not enough than minimum bid");
+            slotBid[slot].initialized = true;
+        }
+        slotWinner[slot] = co;
+        // update bid info for slot
         slotBid[slot].amount = amount;
-		slotBid[slot].sumtotalFees = _sumtotalFees;
-		slotBid[slot].targetProfit = _targetProfit;
-		//update infoSlot
-		infoSlot[slot].maxtargetProfit = _targetProfit;
-		// update burnBid amount
-		//since hubble has no fees currently
-		//sumtotalfees can be 0 
-		//will be removed once hubble has fees
-		//will have to account for sumtotalFees less than _targetProfit but greator than 0
-		if (_sumtotalFees == 0){
-		burnBid = _targetProfit;
-		} else {
-		burnBid = _sumtotalFees - _targetProfit;
-		}
-		//emit event
-		emit currentBestBid(slot, slotBid[slot].amount, slotBid[slot].sumtotalFees, slotBid[slot].targetProfit, co.returnAddress, co.url);
-		return burnBid;
-	}
-	
-	//complete them-done
-	
-	//function by which coordinator bids for himself
-	function bidBySelf(uint32 _slot, string calldata _url, uint _targetProfit, uint _sumtotalFees) external payable {
-	    require(_slot >= currentSlot() + minNextSlots, 'This auction is already closed');
-	    Coordinator memory co = Coordinator(msg.sender, msg.sender, _url);
-		uint burnBid = bid(_slot,co,_targetProfit,_sumtotalFees,msg.value);
-		burnAddress.transfer(burnBid);
-	}
-	
-	//coordinator bids using others address arguments
-	function bidForOthers(
-	    uint32 _slot,
-		uint _targetProfit,
-		uint _sumtotalFees,
-		address payable _returnAddress,
-		address _submitBatchAddress,
-		string calldata _url
-	) external payable {
-	    require(_slot >= currentSlot() + minNextSlots, 'This auction is already closed');
-		Coordinator memory co = Coordinator(_returnAddress, _submitBatchAddress, _url);
-		uint burnBid = bid(_slot,co,_targetProfit,_sumtotalFees,msg.value);
-		burnAddress.transfer(burnBid);
-	}
-	
-	// function needs to be exposed-done
-	/**
+        slotBid[slot].sumtotalFees = _sumtotalFees;
+        slotBid[slot].targetProfit = _targetProfit;
+        //update infoSlot
+        infoSlot[slot].maxtargetProfit = _targetProfit;
+        // update burnBid amount
+        //since hubble has no fees currently
+        //sumtotalfees can be 0
+        //will be removed once hubble has fees
+        //will have to account for sumtotalFees less than _targetProfit but greator than 0 case
+        //ask barry if there should be min txns in block ?-will remove zero sumtotalFees case
+        if (_sumtotalFees == 0) {
+            burnBid = _targetProfit;
+        } else {
+            burnBid = _sumtotalFees - _targetProfit;
+        }
+        //emit event
+        emit currentBestBid(
+            slot,
+            slotBid[slot].amount,
+            slotBid[slot].sumtotalFees,
+            slotBid[slot].targetProfit,
+            co.returnAddress,
+            co.url
+        );
+        return burnBid;
+    }
+
+    //complete them-done
+
+    //function by which coordinator bids for himself
+    function bidBySelf(
+        uint32 _slot,
+        string calldata _url,
+        uint256 _targetProfit,
+        uint256 _sumtotalFees
+    ) external payable {
+        require(
+            _slot >= currentSlot() + minNextSlots,
+            "This auction is already closed"
+        );
+        Coordinator memory co = Coordinator(msg.sender, msg.sender, _url);
+        uint256 burnBid = bid(
+            _slot,
+            co,
+            _targetProfit,
+            _sumtotalFees,
+            msg.value
+        );
+        burnAddress.transfer(burnBid);
+    }
+
+    //coordinator bids using others address arguments
+    function bidForOthers(
+        uint32 _slot,
+        uint256 _targetProfit,
+        uint256 _sumtotalFees,
+        address payable _returnAddress,
+        address _submitBatchAddress,
+        string calldata _url
+    ) external payable {
+        require(
+            _slot >= currentSlot() + minNextSlots,
+            "This auction is already closed"
+        );
+        Coordinator memory co = Coordinator(
+            _returnAddress,
+            _submitBatchAddress,
+            _url
+        );
+        uint256 burnBid = bid(
+            _slot,
+            co,
+            _targetProfit,
+            _sumtotalFees,
+            msg.value
+        );
+        burnAddress.transfer(burnBid);
+    }
+
+    // function needs to be exposed-done
+    /**
      * @dev Retrieve slot winner
      * @return submitBatchAddress,returnAddress,Coordinator url,bidamount
      */
-	function getWinner(uint _slot) external view returns (address, address, string memory, uint) {
-	    address batchSubmitter= slotWinner[_slot].submitBatchAddress;
-		if(batchSubmitter != address(0x00)){ 
-		uint256 amount = slotBid[_slot].amount;
-        address winner = slotWinner[_slot].submitBatchAddress;
-        address beneficiary = slotWinner[_slot].returnAddress;
-        string memory url = slotWinner[_slot].url;
-        return (winner, beneficiary, url, amount);
-		} else {
-		return (coDefault.submitBatchAddress, coDefault.returnAddress, coDefault.url, 0);
-		}
-	}
-	
-	function checkWinner(uint _slot, address _winner) external view returns (bool) {
-	    if (coDefault.submitBatchAddress == _winner) return true;
-		address coordinator = slotWinner[_slot].submitBatchAddress;
-		if(coordinator == _winner){
-			return true;
-		} else {
-			return false;
-		}
-	}
-	
-	//helper functions
-	/**
+    function getWinner(uint256 _slot)
+        external
+        view
+        returns (
+            address,
+            address,
+            string memory,
+            uint256
+        )
+    {
+        address batchSubmitter = slotWinner[_slot].submitBatchAddress;
+        if (batchSubmitter != address(0x00)) {
+            uint256 amount = slotBid[_slot].amount;
+            address winner = slotWinner[_slot].submitBatchAddress;
+            address beneficiary = slotWinner[_slot].returnAddress;
+            string memory url = slotWinner[_slot].url;
+            return (winner, beneficiary, url, amount);
+        } else {
+            return (
+                coDefault.submitBatchAddress,
+                coDefault.returnAddress,
+                coDefault.url,
+                0
+            );
+        }
+    }
+
+    function checkWinner(uint256 _slot, address _winner)
+        external
+        view
+        returns (bool)
+    {
+        if (coDefault.submitBatchAddress == _winner) return true;
+        address coordinator = slotWinner[_slot].submitBatchAddress;
+        if (coordinator == _winner) {
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    //helper functions
+    /**
      * @dev Retrieve block number
      * @return current block number
      */
-    function getBlockNumber() public view returns (uint) {
+    function getBlockNumber() public view returns (uint256) {
         return block.number;
     }
 
@@ -215,12 +284,12 @@ contract BurnAuction {
      * @param numBlock block number
      * @return slot number
      */
-    function block2slot(uint numBlock) public view returns (uint32) {
+    function block2slot(uint256 numBlock) public view returns (uint32) {
         if (numBlock < genesisBlock) return 0;
         return uint32((numBlock - genesisBlock) / (blocksPerSlot));
     }
-	
-	/**
+
+    /**
      * @dev Retrieve current slot
      * @return slot number
      */
@@ -233,7 +302,7 @@ contract BurnAuction {
      * @param slot slot number
      * @return block number
      */
-    function getBlockBySlot(uint32 slot) public view returns (uint) {
-        return (genesisBlock + slot*blocksPerSlot);
+    function getBlockBySlot(uint32 slot) public view returns (uint256) {
+        return (genesisBlock + slot * blocksPerSlot);
     }
 }
